update lighting skill prefabs logic

+++ Assets/Searching/Scripts/ProjectScripts/IDamageable.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public interface IDamageable
 {
     public void TakeDamage(int damage);
-    public void GetStunned(int duration);
+    public void GetStunned(float duration);
 }
